+++ .env
@@ -1,0 +1,3 @@
+POSTGRES_NAME=postgres
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=postgres
