+++ .gitignore
@@ -1,0 +1,42 @@
+# exclude everything into directory
+.devcontainer/
+.vscode
+__pycache__/
+venv/
+
+data/
+
+**/Thumbs.db
+**/thumbs.db
+
+tmp/*
+
+log.txt*
+
+.ddata/
+
+db.sqlite3
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+
+#exclude *.pyc files
+*.pyc
+.idea/
