+++ Dockerfile
@@ -1,0 +1,12 @@
+FROM python:3
+
+WORKDIR /usr/src/app
+
+COPY requirements.txt ./
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+#RUN python manage.py migrate
+
+CMD [ "python", "manage.py", "runserver", "0.0.0.0:8080" ]
