+++ README.md
@@ -1,0 +1,11 @@
+# Automatización de revisión de pdf
+Proyecto para automatizar la revisión de ciertos pdf
+
+# Cómo ejecutar
+
+en Windows, primero asegúrate de tener instalado Docker Desktop, lo puedes descargar de su sitio oficial: https://www.docker.com/products/docker-desktop/ (Es probable que tengas que reiniciar el pc)
+
+Luego ejecutar en la terminal:
+``build.bat`
+
+Y listo!
