--- experimento-pdf3.py
@@ -1,3 +1,0 @@
-import pandas as pd
-
-#hdsjadsa
