+++ src_anterior/experimento-pdf3.py
@@ -1,0 +1,3 @@
+import pandas as pd
+
+#hdsjadsa
